Move checks and actions into subpackages
