+++ app/src/main/scala/com/waz/zclient/security/checks/DeviceAdminCheck.scala
@@ -15,10 +15,11 @@
   * You should have received a copy of the GNU General Public License
   * along with this program.  If not, see <http://www.gnu.org/licenses/>.
   */
-package com.waz.zclient.security
+package com.waz.zclient.security.checks
 
 import android.content.Context
 import com.waz.services.SecurityPolicyService
+import com.waz.zclient.security.SecurityChecklist
 
 import scala.concurrent.Future
